[WFS provider] Split qgswfsfeatureiterator.h/cpp into qgswfsfeaturehitsasyncrequest.h/cpp and qgswfsfeaturedownloaderimpl.h/cpp
No code change. Just better reflect the class organization
